GitHub issue #20: some work to help running the transport on Tomcat
We cannot use old Jetty on our server so I hacked this patch that allows
at least running the transport under Tomcat and apparently even the polling
connection to transport/cometd/connect works. This is enough to silence
the user visible warnings about WFS not working. Not sure if it actually
works but that is not critical for us right now.